LF-4474 Move adding a task and checking for locations out of TaskCount
